cityinfo few tests done

+++ src/main/java/facades/CityInfoFacade.java
@@ -39,7 +39,7 @@ public class CityInfoFacade {
         }
     }
 
-    public CityInfoDTO getCityById(int id) {
+    public CityInfoDTO getCityById(long id) {
         EntityManager em = emf.createEntityManager();
         try {
             CityInfo cityInfo = em.find(CityInfo.class, id);
